use explicit 64 bit ints

+++ structures/bitboard.h
@@ -1,9 +1,10 @@
 #ifndef _BITBOARD_H
 #define _BITBOARD_H
 
+#include <stdint.h>
 #include <stdio.h>
 
-typedef unsigned long long int bitboard_t;
+typedef uint64_t bitboard_t;
 
 #define REMOVE_FIRST_BIT_SET(x) ((x) = ((x)&((x)-1)))
 #define FIRST_BIT_SET(x) (__builtin_ctzll(x))
